splitting ESP32 shims
