+++ uCNC/src/hal/mcus/esp32/mcu_esp32.c
@@ -57,18 +57,7 @@ volatile uint32_t ic74hc595_i2s_pins;
 #endif
 hw_timer_t *esp32_step_timer;
 
-void esp32_usb_wifi_bt_init(void);
-void esp32_usb_wifi_bt_flush(uint8_t *buffer);
-void esp32_usb_wifi_bt_process(void);
-
-#ifdef USE_ARDUINO_SPI_LIBRARY
-#ifdef MCU_HAS_SPI
-void mcu_spi_init(void);
-#endif
-#ifdef MCU_HAS_SPI2
-void mcu_spi2_init(void);
-#endif
-#endif
+#include "../esp32common/esp32_common.h"
 
 #if !defined(RAM_ONLY_SETTINGS) && !defined(USE_ARDUINO_EEPROM_LIBRARY)
 #include <nvs.h>
@@ -83,8 +72,6 @@ typedef struct
 } flash_eeprom_t;
 
 static flash_eeprom_t mcu_eeprom;
-#elif !defined(RAM_ONLY_SETTINGS)
-extern void esp32_eeprom_init(int size);
 #endif
 
 MCU_CALLBACK void mcu_itp_isr(void *arg);
@@ -517,16 +504,30 @@ uint8_t mcu_softpwm_freq_config(uint16_t freq)
 }
 #endif
 
+void mcu_core0_dotasks(void *arg)
+{
+	// loop through received data
+	mcu_uart_dotasks();
+	esp_task_wdt_reset();
+	mcu_uart2_dotasks();
+	esp_task_wdt_reset();
+	mcu_usb_dotasks();
+	esp_task_wdt_reset();
+	mcu_wifi_dotasks();
+	esp_task_wdt_reset();
+	mcu_bt_dotasks();
+}
+
 void mcu_core0_tasks_init(void *arg)
 {
-#ifdef MCU_HAS_UART
-	// install UART driver handler
-	uart_driver_install(UART_PORT, RX_BUFFER_CAPACITY * 2, 0, 0, NULL, 0);
-#endif
-#ifdef MCU_HAS_UART2
-	// install UART driver handler
-	uart_driver_install(UART2_PORT, RX_BUFFER_CAPACITY * 2, 0, 0, NULL, 0);
-#endif
+	mcu_uart_init();
+	mcu_uart_start();
+	mcu_uart2_init();
+	mcu_uart2_start();
+	mcu_wifi_init();
+	mcu_bt_init();
+
+	// xTaskCreatePinnedToCore(mcu_core0_dotasks, "core0Task", 8192, NULL, 7, NULL, 0);
 }
 
 void mcu_rtc_task(void *arg)
@@ -592,21 +593,42 @@ void mcu_init(void)
 	gpio_install_isr_service(0);
 #endif
 
+	/**
+	 * IO conficuration
+	 */
+
 	mcu_io_init();
 
-#ifdef MCU_HAS_UART2
-	// initialize UART
-	const uart_config_t uart2config = {
-			.baud_rate = BAUDRATE2,
-			.data_bits = UART_DATA_8_BITS,
-			.parity = UART_PARITY_DISABLE,
-			.stop_bits = UART_STOP_BITS_1,
-			.flow_ctrl = UART_HW_FLOWCTRL_DISABLE,
-			.source_clk = UART_SCLK_APB};
-	// We won't use a buffer for sending data.
-	uart_param_config(UART2_PORT, &uart2config);
-	uart_set_pin(UART2_PORT, TX2_BIT, RX2_BIT, UART_PIN_NO_CHANGE, UART_PIN_NO_CHANGE);
+#ifdef MCU_HAS_SPI
+	spi_config_t spi_conf = {0};
+	spi_conf.mode = SPI_MODE;
+	mcu_spi_init();
+	mcu_spi_config(spi_conf, SPI_FREQ);
 #endif
+
+#ifdef MCU_HAS_SPI2
+	spi_config_t spi2_conf = {0};
+	spi2_conf.mode = SPI2_MODE;
+	mcu_spi2_init();
+	mcu_spi2_config(spi2_conf, SPI2_FREQ);
+#endif
+
+#ifdef MCU_HAS_I2C
+	mcu_i2c_config(I2C_FREQ);
+#endif
+
+	/**
+	 * EEPROM config
+	 */
+
+	// starts EEPROM before UART to enable WiFi and BT settings
+#if !defined(RAM_ONLY_SETTINGS)
+	mcu_eeprom_init(NVM_STORAGE_SIZE);
+#endif
+
+	/**
+	 * Communications config
+	 */
 
 	// launches isr tasks that will run on core 0
 	// currently it's running PWM and UART on core 0
@@ -614,46 +636,9 @@ void mcu_init(void)
 	// Arduino WiFi ???
 	esp_ipc_call_blocking(0, mcu_core0_tasks_init, NULL);
 
-	// starts EEPROM before UART to enable WiFi and BT settings
-#if !defined(RAM_ONLY_SETTINGS) && !defined(USE_ARDUINO_EEPROM_LIBRARY)
-	// esp32_eeprom_init(NVM_STORAGE_SIZE); // 1K Emulated EEPROM
-
-	// starts nvs
-	mcu_eeprom.size = 0;
-	memset(mcu_eeprom.data, 0, NVM_STORAGE_SIZE);
-	if (nvs_open("eeprom", NVS_READWRITE, &mcu_eeprom.nvs_handle) == ESP_OK)
-	{
-		// determines the maximum sector size of NVS that can be read/write
-		nvs_get_blob(mcu_eeprom.nvs_handle, "eeprom", NULL, &mcu_eeprom.size);
-		if (NVM_STORAGE_SIZE > mcu_eeprom.size)
-		{
-			log_e("eeprom does not have enough space");
-			mcu_eeprom.size = 0;
-		}
-
-		nvs_get_blob(mcu_eeprom.nvs_handle, "eeprom", mcu_eeprom.data, &mcu_eeprom.size);
-	}
-	else
-	{
-		log_e("eeprom failed to open");
-	}
-#elif !defined(RAM_ONLY_SETTINGS)
-	esp32_eeprom_init(NVM_STORAGE_SIZE);
-#endif
-
-#ifdef MCU_HAS_UART
-	// initialize UART
-	const uart_config_t uartconfig = {
-			.baud_rate = BAUDRATE,
-			.data_bits = UART_DATA_8_BITS,
-			.parity = UART_PARITY_DISABLE,
-			.stop_bits = UART_STOP_BITS_1,
-			.flow_ctrl = UART_HW_FLOWCTRL_DISABLE,
-			.source_clk = UART_SCLK_APB};
-	// We won't use a buffer for sending data.
-	uart_param_config(UART_PORT, &uartconfig);
-	uart_set_pin(UART_PORT, TX_BIT, RX_BIT, UART_PIN_NO_CHANGE, UART_PIN_NO_CHANGE);
-#endif
+	/**
+	 * Timers config
+	 */
 
 	// inititialize ITP timer
 	timer_config_t itpconfig = {0};
@@ -681,33 +666,6 @@ void mcu_init(void)
 	// initialize rtc timer (currently on core 1)
 	xTaskCreatePinnedToCore(mcu_rtc_task, "rtcTask", 2048, NULL, 7, NULL, CONFIG_ARDUINO_RUNNING_CORE);
 
-#ifdef MCU_HAS_SPI
-
-	spi_config_t spi_conf = {0};
-#ifndef USE_ARDUINO_SPI_LIBRARY
-	spi_conf.mode = SPI_MODE;
-#else
-	mcu_spi_init();
-#endif
-	mcu_spi_config(spi_conf, SPI_FREQ);
-#endif
-
-#ifdef MCU_HAS_SPI2
-
-	spi_config_t spi2_conf = {0};
-#ifndef USE_ARDUINO_SPI_LIBRARY
-	spi2_conf.mode = SPI2_MODE;
-#else
-	mcu_spi2_init();
-#endif
-	mcu_spi2_config(spi2_conf, SPI2_FREQ);
-#endif
-
-#ifdef MCU_HAS_I2C
-	mcu_i2c_config(I2C_FREQ);
-#endif
-
-	esp32_usb_wifi_bt_init();
 	mcu_enable_global_isr();
 }
 
@@ -760,105 +718,6 @@ void mcu_set_pwm(uint8_t pwm, uint8_t value)
 uint8_t mcu_get_pwm(uint8_t pwm)
 {
 	return 0;
-}
-#endif
-
-/*UART*/
-
-/**
- * sends a uint8_t either via uart (hardware, software or USB virtual COM port)
- * can be defined either as a function or a macro call
- * */
-#ifdef MCU_HAS_UART
-#ifndef UART_TX_BUFFER_SIZE
-#define UART_TX_BUFFER_SIZE 64
-#endif
-DECL_BUFFER(uint8_t, uart_rx, RX_BUFFER_SIZE);
-DECL_BUFFER(uint8_t, uart_tx, UART_TX_BUFFER_SIZE);
-uint8_t mcu_uart_getc(void)
-{
-	uint8_t c = 0;
-	BUFFER_DEQUEUE(uart_rx, &c);
-	return c;
-}
-
-uint8_t mcu_uart_available(void)
-{
-	return BUFFER_READ_AVAILABLE(uart_rx);
-}
-
-void mcu_uart_clear(void)
-{
-	BUFFER_CLEAR(uart_rx);
-}
-
-void mcu_uart_putc(uint8_t c)
-{
-	while (BUFFER_FULL(uart_tx))
-	{
-		mcu_uart_flush();
-	}
-	BUFFER_ENQUEUE(uart_tx, &c);
-}
-
-void mcu_uart_flush(void)
-{
-	while (!BUFFER_EMPTY(uart_tx))
-	{
-		uint8_t tmp[UART_TX_BUFFER_SIZE + 1];
-		memset(tmp, 0, sizeof(tmp));
-		uint8_t r;
-
-		BUFFER_READ(uart_tx, tmp, UART_TX_BUFFER_SIZE, r);
-		uart_write_bytes(UART_PORT, tmp, r);
-	}
-}
-#endif
-
-#ifdef MCU_HAS_UART2
-#ifndef UART2_TX_BUFFER_SIZE
-#define UART2_TX_BUFFER_SIZE 64
-#endif
-DECL_BUFFER(uint8_t, uart2_rx, RX_BUFFER_SIZE);
-DECL_BUFFER(uint8_t, uart2_tx, UART2_TX_BUFFER_SIZE);
-
-uint8_t mcu_uart2_getc(void)
-{
-	uint8_t c = 0;
-	BUFFER_DEQUEUE(uart2_rx, &c);
-	return c;
-}
-
-uint8_t mcu_uart2_available(void)
-{
-	return BUFFER_READ_AVAILABLE(uart2_rx);
-}
-
-void mcu_uart2_clear(void)
-{
-	BUFFER_CLEAR(uart2_rx);
-}
-
-void mcu_uart2_putc(uint8_t c)
-{
-	while (BUFFER_FULL(uart2_tx))
-	{
-		mcu_uart2_flush();
-	}
-	BUFFER_ENQUEUE(uart2_tx, &c);
-}
-
-void mcu_uart2_flush(void)
-{
-	while (!BUFFER_EMPTY(uart2_tx))
-	{
-		uint8_t tmp[UART2_TX_BUFFER_SIZE + 1];
-		memset(tmp, 0, sizeof(tmp));
-		uint8_t r;
-
-		BUFFER_READ(uart2_tx, tmp, UART2_TX_BUFFER_SIZE, r);
-		uart_write_bytes(UART2_PORT, tmp, r);
-	}
 }
 #endif
 
@@ -1000,119 +859,25 @@ void esp32_delay_us(uint16_t delay)
  * */
 void mcu_dotasks(void)
 {
-	// reset WDT
+	// // reset WDT
+	// esp_task_wdt_reset();
+
+	// // loop through received data
+	// mcu_uart_dotasks();
+	// mcu_uart2_dotasks();
+	// mcu_usb_dotasks();
+	// mcu_wifi_dotasks();
+	// mcu_bt_dotasks();
+	mcu_uart_dotasks();
 	esp_task_wdt_reset();
-
-	// loop through received data
-	uint8_t rxdata[RX_BUFFER_SIZE];
-	int rxlen, i;
-#ifdef MCU_HAS_UART
-	rxlen = uart_read_bytes(UART_PORT, rxdata, RX_BUFFER_CAPACITY, 0);
-	for (i = 0; i < rxlen; i++)
-	{
-		uint8_t c = (uint8_t)rxdata[i];
-		if (mcu_com_rx_cb(c))
-		{
-			if (BUFFER_FULL(uart_rx))
-			{
-				STREAM_OVF(c);
-			}
-
-			BUFFER_ENQUEUE(uart_rx, &c);
-		}
-	}
-#endif
-#if defined(MCU_HAS_UART2)
-	rxlen = uart_read_bytes(UART2_PORT, rxdata, RX_BUFFER_CAPACITY, 0);
-	for (i = 0; i < rxlen; i++)
-	{
-		uint8_t c = (uint8_t)rxdata[i];
-#if !defined(DETACH_UART2_FROM_MAIN_PROTOCOL)
-		if (mcu_com_rx_cb(c))
-		{
-			if (BUFFER_FULL(uart2_rx))
-			{
-				STREAM_OVF(c);
-			}
-
-			BUFFER_ENQUEUE(uart2_rx, &c);
-		}
-#else
-		mcu_uart2_rx_cb(c);
-#ifndef UART2_DISABLE_BUFFER
-		if (BUFFER_FULL(uart2_rx))
-		{
-			STREAM_OVF(c);
-		}
-
-		BUFFER_ENQUEUE(uart2_rx, &c);
-#endif
-#endif
-	}
-#endif
-
-	esp32_usb_wifi_bt_process();
+	mcu_uart2_dotasks();
+	esp_task_wdt_reset();
+	mcu_usb_dotasks();
+	esp_task_wdt_reset();
+	mcu_wifi_dotasks();
+	esp_task_wdt_reset();
+	mcu_bt_dotasks();
 }
-
-// Non volatile memory
-/**
- * gets a byte at the given EEPROM (or other non volatile memory) address of the MCU.
- * */
-#if !defined(RAM_ONLY_SETTINGS) && !defined(USE_ARDUINO_EEPROM_LIBRARY)
-uint8_t mcu_eeprom_getc(uint16_t address)
-{
-	if (NVM_STORAGE_SIZE <= address)
-	{
-		DBGMSG("EEPROM invalid address @ %u", address);
-		return 0;
-	}
-#ifndef RAM_ONLY_SETTINGS
-	// return esp32_eeprom_read(address);
-	size_t size = mcu_eeprom.size;
-	if (size)
-	{
-		return mcu_eeprom.data[address];
-	}
-#endif
-	return 0;
-}
-
-/**
- * sets a byte at the given EEPROM (or other non volatile memory) address of the MCU.
- * */
-void mcu_eeprom_putc(uint16_t address, uint8_t value)
-{
-	if (NVM_STORAGE_SIZE <= address)
-	{
-		DBGMSG("EEPROM invalid address @ %u", address);
-	}
-#ifndef RAM_ONLY_SETTINGS
-	// esp32_eeprom_write(address, value);
-	size_t size = mcu_eeprom.size;
-	if (size)
-	{
-		mcu_eeprom.dirty |= (mcu_eeprom.data[address] != value);
-		mcu_eeprom.data[address] = value;
-	}
-#endif
-}
-
-/**
- * flushes all recorded registers into the eeprom.
- * */
-void mcu_eeprom_flush(void)
-{
-#ifndef RAM_ONLY_SETTINGS
-	// esp32_eeprom_flush();
-	// esp32_eeprom_write(address, value);
-	if (mcu_eeprom.size && mcu_eeprom.dirty)
-	{
-		nvs_set_blob(mcu_eeprom.nvs_handle, "eeprom", mcu_eeprom.data, mcu_eeprom.size);
-		nvs_commit(mcu_eeprom.nvs_handle);
-	}
-#endif
-}
-#endif
 
 #ifdef MCU_HAS_ONESHOT_TIMER
 
@@ -1168,364 +933,6 @@ MCU_CALLBACK void mcu_start_timeout()
 #elif defined(MCU_HAS_ONESHOT_TIMER)
 	timer_start(ONESHOT_TIMER_TG, ONESHOT_TIMER_IDX);
 #endif
-}
-#endif
-#endif
-
-#if (defined(MCU_HAS_SPI) && !defined(USE_ARDUINO_SPI_LIBRARY))
-
-static spi_device_handle_t mcu_spi_handle = NULL;
-
-#ifndef mcu_spi_xmit
-uint8_t mcu_spi_xmit(uint8_t data)
-{
-	uint8_t rxdata = 0xFF;
-	spi_transaction_t spi_trans = {0};
-	spi_trans.length = 8; // Number of bits NOT number of bytes.
-	spi_trans.tx_buffer = &data;
-	spi_trans.rx_buffer = &rxdata;
-
-	spi_device_transmit(mcu_spi_handle, &spi_trans);
-
-	return rxdata;
-}
-#endif
-
-#ifndef mcu_spi_config
-void mcu_spi_config(spi_config_t config, uint32_t frequency)
-{
-	mcu_spi_start(config, frequency);
-	mcu_spi_stop();
-}
-#endif
-
-#ifndef mcu_spi_start
-void mcu_spi_start(spi_config_t config, uint32_t frequency)
-{
-	if (mcu_spi_handle)
-	{
-		spi_device_acquire_bus(mcu_spi_handle, portMAX_DELAY);
-		spi_device_release_bus(mcu_spi_handle);
-		spi_bus_remove_device(mcu_spi_handle);
-		mcu_spi_handle = NULL;
-		spi_bus_free(SPI_PORT);
-	}
-
-	spi_bus_config_t spiconf = {
-			.miso_io_num = SPI_SDI_BIT,
-			.mosi_io_num = SPI_SDO_BIT,
-			.sclk_io_num = SPI_CLK_BIT,
-			.quadwp_io_num = -1,
-			.quadhd_io_num = -1,
-			.data4_io_num = -1,
-			.data5_io_num = -1,
-			.data6_io_num = -1,
-			.data7_io_num = -1,
-			.max_transfer_sz = (config.enable_dma) ? SPI_DMA_BUFFER_SIZE : SOC_SPI_MAXIMUM_BUFFER_SIZE,
-			.flags = 0,
-			.intr_flags = 0};
-	// Initialize the SPI bus
-	spi_bus_initialize(SPI_PORT, &spiconf, (config.enable_dma) ? SPI_DMA_CH_AUTO : SPI_DMA_DISABLED);
-
-	spi_device_interface_config_t mcu_spi_conf = {0};
-	mcu_spi_conf.mode = config.mode;
-	mcu_spi_conf.clock_speed_hz = frequency;
-	mcu_spi_conf.spics_io_num = -1;
-	mcu_spi_conf.queue_size = 1;
-	if (config.enable_dma)
-	{
-		mcu_spi_conf.queue_size = 1;
-		mcu_spi_conf.pre_cb = NULL;
-		mcu_spi_conf.pre_cb = NULL;
-	}
-
-	spi_dma_enabled = config.enable_dma;
-	spi_bus_add_device(SPI_PORT, &mcu_spi_conf, &mcu_spi_handle);
-	spi_device_acquire_bus(mcu_spi_handle, portMAX_DELAY);
-}
-#endif
-
-#ifndef mcu_spi_stop
-void mcu_spi_stop(void)
-{
-	spi_device_release_bus(mcu_spi_handle);
-	mcu_spi_handle = NULL;
-}
-#endif
-
-#ifndef mcu_spi_bulk_transfer
-// data buffer for normal or DMA
-static FORCEINLINE bool mcu_spi_transaction(const uint8_t *out, uint8_t *in, uint16_t len)
-{
-	static void *o = NULL;
-	static void *i = NULL;
-	static bool is_running = false;
-	static spi_transaction_t t = {0};
-
-	// start a new transmition
-	if (!is_running)
-	{
-		if (spi_dma_enabled)
-		{
-			// DMA requires 4byte aligned transfers
-			uint16_t l = (((len >> 2) + ((len & 0x03) ? 1 : 0)) << 2);
-			o = heap_caps_malloc(l, MALLOC_CAP_DMA);
-			memcpy(o, out, len);
-			if (in)
-			{
-				i = heap_caps_malloc(l, MALLOC_CAP_DMA);
-				memset(i, 0x00, l);
-			}
-		}
-		else
-		{
-			o = out;
-			if (in)
-			{
-				i = in;
-			}
-		}
-
-		t.length = len * 8; // Length in bits
-		t.tx_buffer = o;
-		t.rxlength = 0; // this deafults to length
-
-		if (in)
-		{
-			t.rx_buffer = i;
-		}
-
-		spi_device_polling_start(mcu_spi_handle, &t, portMAX_DELAY);
-		is_running = true;
-	}
-	else
-	{
-		// check transfer state
-		if (spi_device_polling_end(mcu_spi_handle, 0) == ESP_OK)
-		{
-			if (spi_dma_enabled)
-			{
-				// copy back memory from DMA
-				if (in)
-				{
-					memcpy(in, i, len);
-					heap_caps_free(i);
-				}
-				heap_caps_free(o);
-			}
-
-			is_running = false;
-			return false;
-		}
-	}
-
-	return true;
-}
-
-bool mcu_spi_bulk_transfer(const uint8_t *out, uint8_t *in, uint16_t len)
-{
-	static uint16_t data_offset = 0;
-	uint32_t offset = data_offset;
-	uint16_t max_transfer_size = (spi_dma_enabled) ? SPI_DMA_BUFFER_SIZE : SOC_SPI_MAXIMUM_BUFFER_SIZE;
-
-	uint8_t *i = NULL;
-	if (in)
-	{
-		i = &in[offset];
-	}
-
-	if (!mcu_spi_transaction(&out[offset], i, MIN(max_transfer_size, (len - offset))))
-	{
-		offset += max_transfer_size;
-		if (offset >= len)
-		{
-			data_offset = 0;
-			return false;
-		}
-
-		data_offset = (uint16_t)offset;
-	}
-
-	return true;
-}
-#endif
-#endif
-
-#if (defined(MCU_HAS_SPI2) && !defined(USE_ARDUINO_SPI_LIBRARY))
-
-static spi_device_handle_t mcu_spi2_handle = NULL;
-
-#ifndef mcu_spi2_xmit
-uint8_t mcu_spi2_xmit(uint8_t data)
-{
-	uint8_t rxdata = 0xFF;
-	spi_transaction_t spi_trans = {0};
-	spi_trans.length = 8; // Number of bits NOT number of bytes.
-	spi_trans.tx_buffer = &data;
-	spi_trans.rx_buffer = &rxdata;
-
-	spi_device_transmit(mcu_spi2_handle, &spi_trans);
-
-	return rxdata;
-}
-#endif
-
-#ifndef mcu_spi2_config
-void mcu_spi2_config(spi_config_t config, uint32_t frequency)
-{
-	mcu_spi2_start(config, frequency);
-	mcu_spi2_stop();
-}
-#endif
-
-#ifndef mcu_spi2_start
-void mcu_spi2_start(spi_config_t config, uint32_t frequency)
-{
-	if (mcu_spi2_handle)
-	{
-		spi_device_acquire_bus(mcu_spi2_handle, portMAX_DELAY);
-		spi_device_release_bus(mcu_spi2_handle);
-		spi_bus_remove_device(mcu_spi2_handle);
-		mcu_spi2_handle = NULL;
-		spi_bus_free(SPI2_PORT);
-	}
-
-	spi_bus_config_t spiconf = {
-			.miso_io_num = SPI2_SDI_BIT,
-			.mosi_io_num = SPI2_SDO_BIT,
-			.sclk_io_num = SPI2_CLK_BIT,
-			.quadwp_io_num = -1,
-			.quadhd_io_num = -1,
-			.data4_io_num = -1,
-			.data5_io_num = -1,
-			.data6_io_num = -1,
-			.data7_io_num = -1,
-			.max_transfer_sz = (config.enable_dma) ? SPI2_DMA_BUFFER_SIZE : SOC_SPI_MAXIMUM_BUFFER_SIZE,
-			.flags = 0,
-			.intr_flags = 0};
-	// Initialize the SPI2 bus
-	spi_bus_initialize(SPI2_PORT, &spiconf, (config.enable_dma) ? SPI_DMA_CH_AUTO : SPI_DMA_DISABLED);
-
-	spi_device_interface_config_t mcu_spi_conf = {0};
-	mcu_spi_conf.mode = config.mode;
-	mcu_spi_conf.clock_speed_hz = frequency;
-	mcu_spi_conf.spics_io_num = -1;
-	mcu_spi_conf.queue_size = 1;
-	if (config.enable_dma)
-	{
-		mcu_spi_conf.queue_size = 1;
-		mcu_spi_conf.pre_cb = NULL;
-		mcu_spi_conf.pre_cb = NULL;
-	}
-
-	spi2_dma_enabled = config.enable_dma;
-	spi_bus_add_device(SPI2_PORT, &mcu_spi_conf, &mcu_spi2_handle);
-	spi_device_acquire_bus(mcu_spi2_handle, portMAX_DELAY);
-}
-#endif
-
-#ifndef mcu_spi2_stop
-void mcu_spi2_stop(void)
-{
-	spi_device_release_bus(mcu_spi2_handle);
-	mcu_spi2_handle = NULL;
-}
-#endif
-
-#ifndef mcu_spi2_bulk_transfer
-// data buffer for normal or DMA
-static FORCEINLINE bool mcu_spi2_transaction(const uint8_t *out, uint8_t *in, uint16_t len)
-{
-	static void *o = NULL;
-	static void *i = NULL;
-	static bool is_running = false;
-	static spi_transaction_t t = {0};
-
-	// start a new transmition
-	if (!is_running)
-	{
-		if (spi2_dma_enabled)
-		{
-			// DMA requires 4byte aligned transfers
-			uint16_t l = (((len >> 2) + ((len & 0x03) ? 1 : 0)) << 2);
-			o = heap_caps_malloc(l, MALLOC_CAP_DMA);
-			memcpy(o, out, len);
-			if (in)
-			{
-				i = heap_caps_malloc(l, MALLOC_CAP_DMA);
-				memset(i, 0x00, l);
-			}
-		}
-		else
-		{
-			o = out;
-			if (in)
-			{
-				i = in;
-			}
-		}
-
-		t.length = len * 8; // Length in bits
-		t.tx_buffer = o;
-		t.rxlength = 0; // this deafults to length
-
-		if (in)
-		{
-			t.rx_buffer = i;
-		}
-
-		spi_device_polling_start(mcu_spi2_handle, &t, portMAX_DELAY);
-		is_running = true;
-	}
-	else
-	{
-		// check transfer state
-		if (spi_device_polling_end(mcu_spi2_handle, 0) == ESP_OK)
-		{
-			if (spi2_dma_enabled)
-			{
-				// copy back memory from DMA
-				if (in)
-				{
-					memcpy(in, i, len);
-					heap_caps_free(i);
-				}
-				heap_caps_free(o);
-			}
-
-			is_running = false;
-			return false;
-		}
-	}
-
-	return true;
-}
-
-bool mcu_spi2_bulk_transfer(const uint8_t *out, uint8_t *in, uint16_t len)
-{
-	static uint16_t data_offset = 0;
-	uint32_t offset = data_offset;
-	uint16_t max_transfer_size = (spi2_dma_enabled) ? SPI2_DMA_BUFFER_SIZE : SOC_SPI_MAXIMUM_BUFFER_SIZE;
-
-	uint8_t *i = NULL;
-	if (in)
-	{
-		i = &in[offset];
-	}
-
-	if (!mcu_spi2_transaction(&out[offset], i, MIN(max_transfer_size, (len - offset))))
-	{
-		offset += max_transfer_size;
-		if (offset >= len)
-		{
-			data_offset = 0;
-			return false;
-		}
-
-		data_offset = (uint16_t)offset;
-	}
-
-	return true;
 }
 #endif
 #endif
