+++ uCNC/src/hal/mcus/esp32common/esp32_i2c.cpp
@@ -1,0 +1,77 @@
+/*
+	Name: esp32_arduino.cpp
+	Description: Contains all Arduino ESP32 C++ to C functions used by µCNC.
+
+	Copyright: Copyright (c) João Martins
+	Author: João Martins
+	Date: 27-07-2022
+
+	µCNC is free software: you can redistribute it and/or modify
+	it under the terms of the GNU General Public License as published by
+	the Free Software Foundation, either version 3 of the License, or
+	(at your option) any later version. Please see <http://www.gnu.org/licenses/>
+
+	µCNC is distributed WITHOUT ANY WARRANTY;
+	Also without the implied warranty of	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+	See the	GNU General Public License for more details.
+*/
+
+#if defined(ESP32) /*&& !(defined(ESP32C3) || defined(ESP32S3))*/
+#include <Arduino.h>
+#include <Wire.h>
+
+extern "C"
+{
+#include "../../../cnc.h"
+
+#ifdef MCU_HAS_I2C
+
+#if (I2C_ADDRESS != 0)
+	static uint8_t mcu_i2c_buffer_len;
+	static uint8_t mcu_i2c_buffer[I2C_SLAVE_BUFFER_SIZE];
+	void esp32_i2c_onreceive(int len)
+	{
+		uint8_t l = I2C_REG.readBytes(mcu_i2c_buffer, len);
+		mcu_i2c_slave_cb(mcu_i2c_buffer, &l);
+		mcu_i2c_buffer_len = l;
+	}
+
+	void esp32_i2c_onrequest(void)
+	{
+		I2C_REG.write(mcu_i2c_buffer, mcu_i2c_buffer_len);
+	}
+
+#endif
+
+	void mcu_i2c_config(uint32_t frequency)
+	{
+#if (I2C_ADDRESS == 0)
+		I2C_REG.begin(I2C_DATA_BIT, I2C_CLK_BIT, frequency);
+#else
+		I2C_REG.onReceive(esp32_i2c_onreceive);
+		I2C_REG.onRequest(esp32_i2c_onrequest);
+		I2C_REG.begin(I2C_ADDRESS, I2C_DATA_BIT, I2C_CLK_BIT, frequency);
+#endif
+	}
+
+	uint8_t mcu_i2c_send(uint8_t address, uint8_t *data, uint8_t datalen, bool release, uint32_t ms_timeout)
+	{
+		I2C_REG.beginTransmission(address);
+		I2C_REG.write(data, datalen);
+		return (I2C_REG.endTransmission(release) == 0) ? I2C_OK : I2C_NOTOK;
+	}
+
+	uint8_t mcu_i2c_receive(uint8_t address, uint8_t *data, uint8_t datalen, uint32_t ms_timeout)
+	{
+		I2C_REG.setTimeOut((uint16_t)ms_timeout);
+		if (I2C_REG.requestFrom(address, datalen) == datalen)
+		{
+			I2C_REG.readBytes(data, datalen);
+			return I2C_OK;
+		}
+
+		return I2C_NOTOK;
+	}
+#endif
+}
+#endif
