+++ uCNC/src/hal/mcus/esp32s3/mcu_esp32s3.c
@@ -84,7 +84,7 @@ static volatile bool esp32_global_isr_enabled;
 
 // static flash_eeprom_t mcu_eeprom;
 // #elif !defined(RAM_ONLY_SETTINGS)
-// extern void esp32_eeprom_init(int size);
+// extern void mcu_eeprom_init(int size);
 // #endif
 
 // MCU_CALLBACK void mcu_itp_isr(void *arg);
@@ -611,9 +611,9 @@ void mcu_init(void)
 
 #if !defined(RAM_ONLY_SETTINGS)
 #ifdef USE_ARDUINO_EEPROM_LIBRARY
-	extern void esp32_eeprom_init(int size);
+	extern void mcu_eeprom_init(int size);
 #endif
-	esp32_eeprom_init(NVM_STORAGE_SIZE);
+	mcu_eeprom_init(NVM_STORAGE_SIZE);
 #endif
 
 	xTaskCreatePinnedToCore(mcu_rtc_task, "rtcTask", 4096, NULL, 7, NULL, CONFIG_ARDUINO_RUNNING_CORE);
@@ -640,7 +640,7 @@ void mcu_init(void)
 
 	// 	// starts EEPROM before UART to enable WiFi and BT settings
 	// #if !defined(RAM_ONLY_SETTINGS) && !defined(USE_ARDUINO_EEPROM_LIBRARY)
-	// 	// esp32_eeprom_init(NVM_STORAGE_SIZE); // 1K Emulated EEPROM
+	// 	// mcu_eeprom_init(NVM_STORAGE_SIZE); // 1K Emulated EEPROM
 
 	// 	// starts nvs
 	// 	mcu_eeprom.size = 0;
@@ -662,7 +662,7 @@ void mcu_init(void)
 	// 		log_e("eeprom failed to open");
 	// 	}
 	// #elif !defined(RAM_ONLY_SETTINGS)
-	// 	esp32_eeprom_init(NVM_STORAGE_SIZE);
+	// 	mcu_eeprom_init(NVM_STORAGE_SIZE);
 	// #endif
 
 	// #ifdef MCU_HAS_UART
