+++ uCNC/uCNC.ino.cpp
@@ -1,0 +1,21 @@
+# 1 "C:\\Users\\JCEM\\AppData\\Local\\Temp\\tmpw6fub026"
+#include <Arduino.h>
+# 1 "C:/Users/JCEM/Documents/GitHub/uCNC/uCNC/uCNC.ino"
+#include "src/cnc.h"
+void setup();
+void loop();
+#line 3 "C:/Users/JCEM/Documents/GitHub/uCNC/uCNC/uCNC.ino"
+void setup()
+{
+
+ ucnc_init();
+
+
+}
+
+void loop()
+{
+
+ ucnc_run();
+# 24 "C:/Users/JCEM/Documents/GitHub/uCNC/uCNC/uCNC.ino"
+}
